apps/products/views.py
from rest_framework import generics, permissions
from .models import Category, Product, ProductReview, Wishlist
from .serializers import (
    CategorySerializer, ProductSerializer, 
    ProductReviewSerializer, WishlistSerializer
)

class CategoryListView(generics.ListAPIView):
    queryset = Category.objects.filter(parent__isnull=True)
    serializer_class = CategorySerializer

class ProductListView(generics.ListAPIView):
    queryset = Product.objects.prefetch_related('images', 'sizes', 'reviews').all()
    serializer_class = ProductSerializer

class ProductDetailView(generics.RetrieveAPIView):
    queryset = Product.objects.prefetch_related('images', 'sizes', 'reviews').all()
    serializer_class = ProductSerializer
    lookup_field = 'id'

class ReviewCreateView(generics.CreateAPIView):
    queryset = ProductReview.objects.all()
    serializer_class = ProductReviewSerializer
    permission_classes = [permissions.IsAuthenticatedOrReadOnly]

    def perform_create(self, serializer):
        serializer.save(user=self.request.user)

class WishlistListCreateView(generics.ListCreateAPIView):
    serializer_class = WishlistSerializer
    permission_classes = [permissions.IsAuthenticated]

    def get_queryset(self):
        return Wishlist.objects.filter(user=self.request.user)

    def perform_create(self, serializer):
        serializer.save(user=self.request.user)